Implemented a minimal reproduction of issues with core-data use
There is a regularly reproducible issue when altering a checkbox
in the plugin, where after during save of the record a source
I cannot determine is calling `getEntityRecord`, loading the out
of date record via apiFetch and storing it in state causing the
state to be incorrect and the UI to flicker.

Every time I try to lock down this particular call I cannot trace
it for some reason, I always get lost in the bowels of the
framework.